% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pracas}
\alias{pracas}
\title{Dados das praças e canteiros da cidade de São Paulo}
\format{
Um data frame com 6500 linhas e 12 variáveis
}
\usage{
data(pracas)
}
\description{
Um dataset contendo atributos de 6500 praças e canteiros da cidade
de São paulo. As variáveis são as seguintes:
}
\details{
\itemize{
\item ativo. número de identificação das praças ou canteiros
\item descricao. tipo de objeto (Praça, Canteiro Central)
\item logradouro. nome da rua em que fica localizada a praça ou o canteiro
\item numero. complemento numérico do logradouro
\item bairro. bairro em que fica situada a praça ou canteiro
\item cep. código do endereço
\item zona. zona dentro da cidade de São Paulo em que fica localizada a praça ou canteiro (Norte, Sul, Centro, Leste, Oeste, Sem identificação)
\item status. se o cadastro da praça está ativo ou não (Ativo, Inativo)
\item subprefeitura. subprefeitura em que fica localizada a praça ou canteiro
\item latitude. latitude da praça ou canteiro
\item longitude. longitude da praça ou canteiro
\item area. área aproximada da praça ou canteiro em metros quadrados (m2).
}
}
\keyword{datasets}
