% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{termos}
\alias{termos}
\title{Dados dos termos de compromisso relacionado às praças e canteiros da cidade de São Paulo}
\format{
Um data frame com 807 linhas e 4 variáveis
}
\usage{
data(termos)
}
\description{
Um dataset contendo informações dos 807 termos de compromisso relacionados
às praças e canteiros da cidade de São paulo. As variáveis são as seguintes:
}
\details{
\itemize{
\item processo_sei. número SEI do processo administrativo que gerou o termo de compromisso da praça ou canteiro
\item praca_canteiro. nome da praça ou canteiro
\item dt_expiracao. data de expiração do termo de compromisso.
\item subprefeitura. subprefeitura em que fica localizada a praça ou canteiro.
}
}
\keyword{datasets}
